+++ brave_releases_checker/main.py
@@ -182,24 +182,15 @@ class BraveReleaseChecker:  # pylint: disable=R0902,R0903
 
     def _get_installed_version_opensuse(self) -> Union[version.Version, None]:
         """Gets installed version on openSUSE."""
-        try:
-            process = subprocess.run(['zypper', 'info', self.package_name_prefix], capture_output=True, text=True, check=True)
+        process = subprocess.run(['zypper', 'info', self.package_name_prefix], capture_output=True, text=True, check=True)
+        if process.returncode == 0:
             output = process.stdout
             for line in output.splitlines():
-                if line.startswith('Version: '):
-                    version_str = line.split(':')[-1].strip()
+                if line.startswith('Version'):
+                    version_str = line.split(':')[1].split('-')[0].strip()
                     print(f"Installed Package (openSUSE): {self.package_name_prefix} - Version: {version_str}")
                     return version.parse(version_str)
-        except subprocess.CalledProcessError as e:
-            if f"Information for package '{self.package_name_prefix}' not found." in e.stderr:
-                print(f"Package {self.package_name_prefix} is not installed on this openSUSE system.")
-                sys.exit(1)
-            else:
-                print(f"{self.color.bred}Error:{self.color.endc} checking package (openSUSE): {e}")
-                sys.exit(1)
-        except FileNotFoundError:
-            print(f"{self.color.bred}Error:{self.color.endc} zypper command not found.")
-            sys.exit(1)
+        print(f"Package {self.package_name_prefix} is not installed on this openSUSE system.")
         return None
 
     def _fetch_github_releases(self) -> list: